+++ inst/testthat/helper_expectations.R
@@ -380,9 +380,11 @@ expect_learner = function(lrn, task = NULL, check_man = TRUE) {
     checkmate::expect_subset(lrn$properties, mlr3::mlr_reflections$learner_properties[[task$task_type]])
     testthat::expect_identical(lrn$task_type, task$task_type)
 
-    if ("bundle" %in% lrn$properties) {
-      expect_bundleable(lrn, task)
+    if ("marshal" %in% lrn$properties) {
+      expect_marshallable_learner(lrn, task)
     }
+  } else if ("marshal" %in% lrn$properties) {
+    message("Cannot test 'marshal' property of the learner as no task is provided.")
   }
 
   if (!inherits(lrn, "GraphLearner") && !inherits(lrn, "AutoTuner")) { # still not in pipelines, breaking check in mlr3tuning
@@ -391,48 +393,51 @@ expect_learner = function(lrn, task = NULL, check_man = TRUE) {
 
 }
 
-expect_bundleable = function(learner, task) {
-  expect_true("bundle" %in% learner$properties)
+expect_marshallable_learner = function(learner, task) {
+  expect_true("marshal" %in% learner$properties)
   learner$state = NULL
 
   has_public = function(learner, x) {
     exists(x, learner, inherits = FALSE)
   }
 
-  expect_true(has_public(learner, "bundle") && test_function(learner$bundle, nargs = 0))
-  expect_true(has_public(learner, "unbundle") && test_function(learner$unbundle, nargs = 0))
-  expect_true(has_public(learner, "bundle"))
-  expect_true(has_public(learner, "bundle_model") && test_function(learner$bundle_model, nargs = 1, args = "model"))
-  expect_true(has_public(learner, "unbundle_model") && test_function(learner$unbundle_model, nargs = 1, args = "model"))
+  expect_true(has_public(learner, "marshal") && test_function(learner$marshal, nargs = 0))
+  expect_true(has_public(learner, "unmarshal") && test_function(learner$unmarshal, nargs = 0))
+  expect_true(has_public(learner, "marshalled"))
 
-  expect_false(learner$bundled)
-
-  # (un)bundling only possible after training
-  expect_error(learner$bundle(), "has not been trained")
-  expect_error(learner$unbundle(), "has not been trained")
+  # (un)marshal only possible after training
+  expect_error(learner$marshal(), "has not been trained")
+  expect_error(learner$unmarshal(), "has not been trained")
+  expect_error(learner$marshalled, "has not been trained")
 
   learner$train(task)
   model = learner$model
-  expect_false(learner$bundled)
-  learner$bundle()
-  expect_true(learner$bundled)
+  class_prev = class(model)
+  expect_false(learner$marshalled)
+  expect_equal(marshalled_model(learner$model), learner$marshalled)
+  expect_invisible(learner$marshal())
+  expect_true(learner$marshalled)
+  expect_equal(marshalled_model(learner$model), learner$marshalled)
 
-  # cannot predict with bundled learner
-  expect_error(learner$predict(task), "has not been unbundled")
-  expect_true(identical(learner$model, "bundle"))
+  # cannot predict with marshalled learner
+  expect_error(learner$predict(task), "has not been unmarshalled")
 
-  # unbundling works
-  learner$unbundle()
-  # can predict after unbundling
+  # unmarshalling works
+  expect_invisible(learner$unmarshal())
+  # can predict after unmarshalling
   expect_prediction(learner$predict(task))
   # model is reset
   expect_equal(learner$model, model)
-  # bundled is set accordingly
-  expect_false(learner$bundled)
+  # marshalled is set accordingly
+  expect_false(learner$marshalled)
 
-  # when re-training, bundled is reset
+  expect_equal(class(learner$model), class_prev)
+
+  # when re-training, marshalled is reset
   learner$predict(task)
-  expect_false(learner$train(task)$bundled)
+  expect_false(learner$train(task)$marshalled)
+
+
 
 }