+++ man-roxygen/param_bundle.R
@@ -1,3 +1,0 @@
-#' @param bundle (`logical(1)`)\cr
-#' Whether to bundle the learner(s) after the train-predict loop.
-#' Default is `TRUE`.