+++ man-roxygen/param_unmarshal.R
@@ -1,0 +1,2 @@
+#' @param unmarshal [`Learner`]\cr
+#'   Whether to unmarshal the learner.