+++ tests/testthat/test_resample.R
@@ -157,74 +157,64 @@ test_that("as_resample_result works for result data", {
   expect_class(rr2, "ResampleResult")
 })
 
-test_that("bundling", {
-  task = tsk("mtcars")
-  LearnerRegrTest = R6Class("LearnerRegrTest",
-    inherit = LearnerRegrFeatureless,
-    public = list(
-      initialize = function() {
-        super$initialize()
-        self$properties = c("bundle", self$properties)
-      },
-      bundle = function() learner_bundle(self),
-      unbundle = function() learner_unbundle(self),
-      bundle_model = function(model) {
-        structure(list(model), class = "bundled")
-      },
-      unbundle_model = function(model) {
-        model[[1L]]
-      }
-    ),
-    active = list(
-      bundled = function() learner_bundled(self)
-    ),
-    private = list(
-      .tmp_model = NULL
-    )
-  )
-  learner = LearnerRegrTest$new()
-
-  # allow to bundle during resample()
-  resampling = rsmp("cv", folds = 2L)
-  resampling$instantiate(task) # to compare different resample results
-  rr1 = resample(task, learner, resampling, store_models = TRUE, bundle = TRUE)
-  lrn_rec = rr1$learners[[1L]]
-  expect_true(lrn_rec$bundled)
-  expect_false(lrn_rec$unbundle()$bundled)
-
-  # ResampleResult can be unbundled.
-  expect_true(rr1$learners[[1]]$bundled)
-  expect_class(rr1$learners[[1]]$model, "bundled")
-  expect_true(rr1$learners[[2]]$bundled)
-  expect_class(rr1$learners[[2]]$model, "bundled")
-
-  rr1$unbundle()
-  expect_false(rr1$learners[[1]]$bundled)
-  expect_class(rr1$learners[[1]]$model, "regr.featureless_model")
-  expect_false(rr1$learners[[2]]$bundled)
-  expect_class(rr1$learners[[2]]$model, "regr.featureless_model")
-
-  rr1$bundle()
-  expect_true(rr1$learners[[1]]$bundled)
-  expect_class(rr1$learners[[1]]$model, "bundled")
-  expect_true(rr1$learners[[2]]$bundled)
-  expect_class(rr1$learners[[2]]$model, "bundled")
-
-  # bundled resamples are equivalent to unbundled results
-  rr2 = resample(task, lrn("regr.featureless"), resampling, store_models = TRUE)
-  rr3 = resample(task, learner, resampling, store_models = FALSE)
-  expect_equal(rr1$aggregate(), rr3$aggregate())
-  expect_equal(rr1$aggregate(), rr2$aggregate())
-
-  # bundling can be disabled
-  rr4 = resample(task, learner, resampling, bundle = FALSE, store_models = TRUE)
-  expect_false(rr4$learners[[1]]$bundled)
+test_that("parallel execution automatically triggers marshalling", {
+  learner = lrn("classif.lily", count_marshalling = TRUE)
+  task = tsk("iris")
+  resampling = rsmp("holdout")
+  rr = with_future(future::multisession, {
+    resample(task, learner, resampling, store_models = TRUE, unmarshal = TRUE)
+  })
+  expect_equal(rr$learners[[1]]$model$marshal_count, 1)
+  expect_false(rr$learners[[1]]$marshalled)
 })
 
-test_that("bundling does nothing with unbundleable learners", {
-  rr = resample(tsk("mtcars"), lrn("regr.featureless"), rsmp("cv", folds = 2))
-  rr$unbundle()
-  expect_resample_result(rr)
-  rr$bundle()
-  expect_resample_result(rr)
+test_that("sequential execution does not trigger marshalling", {
+  learner = lrn("classif.lily", count_marshalling = TRUE)
+  task = tsk("iris")
+  resampling = rsmp("holdout")
+  rr = with_future(future::sequential, {
+    resample(task, learner, resampling, store_models = TRUE, unmarshal = TRUE)
+  })
+  expect_equal(rr$learners[[1]]$model$marshal_count, 0)
+})
+
+test_that("parallel exeuction and callr marshall twice", {
+  learner = lrn("classif.lily", count_marshalling = TRUE, encapsulate = c(train = "callr"))
+  task = tsk("iris")
+  resampling = rsmp("holdout")
+  rr = with_future(future::multisession, {
+    resample(task, learner, resampling, store_models = TRUE, unmarshal = TRUE)
+  })
+  expect_equal(rr$learners[[1]]$model$marshal_count, 2)
+  expect_false(rr$learners[[1]]$marshalled)
+})
+
+
+test_that("unmarshal parameter is respected", {
+  learner = lrn("classif.lily", count_marshalling = TRUE, encapsulate = c(train = "callr"))
+  task = tsk("iris")
+  resampling = rsmp("holdout")
+  rr = with_future(future::multisession, {
+    list(
+      marshalled = resample(task, learner, resampling, store_models = TRUE, unmarshal = FALSE),
+      unmarshalled = resample(task, learner, resampling, store_models = TRUE, unmarshal = TRUE)
+    )
+  })
+  expect_false(rr$unmarshalled$learners[[1]]$marshalled)
+  expect_true(rr$marshalled$learners[[1]]$marshalled)
+})
+
+test_that("ResampleResult can be (un)marshalled", {
+  rr = resample(tsk("iris"), lrn("classif.lily"), rsmp("holdout"), store_models = TRUE)
+  expect_false(rr$learners[[1]]$marshalled)
+  rr$marshal()
+  expect_true(rr$learners[[1]]$marshalled)
+  rr$unmarshal()
+  expect_false(rr$learners[[1]]$marshalled)
+
+  # also works with non-marshallable learner
+  rr1 = resample(tsk("iris"), lrn("classif.featureless"), rsmp("holdout"), store_models = TRUE)
+  model = rr1$learners[[1]]$model
+  rr1$unmarshal()
+  expect_equal(rr1$learners[[1]]$model, model)
 })