+++ NEWS.md
@@ -1,8 +1,8 @@
 # mlr3 (development version)
 
-* Feat: added support for `"bundle"` property, which allows learners to process
+* Feat: added support for `"marshal"` property, which allows learners to process
 models so they can be serialized. This happens automatically during `resample()`
-and `benchmark()`. The naming was inspired by the {bundle} package.
+and `benchmark()`. The naming was inspired by the {marshal} package.
 
 # mlr3 0.17.2